[FEAT] Choose a specific process or complete process
The following modifications were made: The option to choose a specific process (TSE, Treasury, National Registry, CCSS) or the complete process was added. Added validation to verify if the person is a minor through a checkbox and execute the corresponding flow. The bug related to captures in the TSE process was reviewed and corrected.